fix overlay key event for last overlay
