+++ packages/vlossom/src/models/overlay-model.ts
@@ -1,10 +1,14 @@
 import { PropType } from 'vue';
-import { ColorScheme } from '@/declaration';
+import { ColorScheme, OverlayCallbacks } from '@/declaration';
 
 export function getOverlayProps<T>() {
     return {
         colorScheme: { type: String as PropType<ColorScheme> },
         styleSet: { type: [String, Object] as PropType<string | T> },
+        callbacks: {
+            type: Object as PropType<OverlayCallbacks>,
+            default: () => ({}),
+        },
         dimClose: { type: Boolean, default: true },
         dimmed: { type: Boolean, default: true },
         escClose: { type: Boolean, default: true },
